requirements.txt
asgiref==3.4.1
Django==3.2.6
django-bootstrap-pagination==1.7.1
django-crispy-forms==1.13.0
django-extra-views==0.14.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
isort==5.9.3
Pillow==8.3.2
PyJWT==2.2.0
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
unicode==2.8
Unidecode==1.3.2
